+++ resources/views/posts/show.blade.php
@@ -16,6 +16,7 @@
                 <div class="ob-field">
                     <label for="comment-body">{{ __('openbook.comments.new_label') }}</label>
                     <textarea id="comment-body" name="body" rows="3" required maxlength="{{ config('openbook.comments.max_length', 2000) }}" data-mention-autocomplete>{{ old('body') }}</textarea>
+                    <p class="ob-field__help">{{ __('openbook.composer.markdown_help') }}</p>
                     <div class="ob-emoji-toolbar">
                         <x-emoji-trigger target="comment-body" />
                     </div>